test datatyps

+++ grbl/gcode.h
@@ -143,31 +143,31 @@
 
 // NOTE: When this struct is zeroed, the above defines set the defaults for the system.
 typedef struct {
-  uint8_t motion;          // {G0,G1,G2,G3,G38.2,G80}
-  uint8_t feed_rate;       // {G93,G94}
-  uint8_t units;           // {G20,G21}
-  uint8_t distance;        // {G90,G91}
+  uint32_t motion;          // {G0,G1,G2,G3,G38.2,G80}
+  uint32_t feed_rate;       // {G93,G94}
+  uint32_t units;           // {G20,G21}
+  uint32_t distance;        // {G90,G91}
   // uint8_t distance_arc; // {G91.1} NOTE: Don't track. Only default supported.
-  uint8_t plane_select;    // {G17,G18,G19}
+  uint32_t plane_select;    // {G17,G18,G19}
   // uint8_t cutter_comp;  // {G40} NOTE: Don't track. Only default supported.
-  uint8_t tool_length;     // {G43.1,G49}
-  uint8_t coord_select;    // {G54,G55,G56,G57,G58,G59}
+  uint32_t tool_length;     // {G43.1,G49}
+  uint32_t coord_select;    // {G54,G55,G56,G57,G58,G59}
   // uint8_t control;      // {G61} NOTE: Don't track. Only default supported.
-  uint8_t program_flow;    // {M0,M1,M2,M30}
-  uint8_t coolant;         // {M7,M8,M9}
-  uint8_t spindle;         // {M3,M4,M5}
+  uint32_t program_flow;    // {M0,M1,M2,M30}
+  uint32_t coolant;         // {M7,M8,M9}
+  uint32_t spindle;         // {M3,M4,M5}
 } gc_modal_t;  
 
 typedef struct {
   float f;         // Feed
   float ijk[3];    // I,J,K Axis arc offsets
-  uint8_t l;       // G10 or canned cycles parameters
+  uint32_t l;       // G10 or canned cycles parameters
   int32_t n;       // Line number
   float p;         // G10 or dwell parameters
   // float q;      // G82 peck drilling
   float r;         // Arc radius
   float s;         // Spindle speed
-  uint8_t t;       // Tool selection
+  uint32_t t;       // Tool selection
   float xyz[N_AXIS];    // X,Y,Z Translational axes
 } gc_values_t;
 
@@ -177,7 +177,7 @@ typedef struct {
   
   float spindle_speed;          // RPM
   float feed_rate;              // Millimeters/min
-  uint8_t tool;                 // Tracks tool number. NOT USED.
+  uint32_t tool;                 // Tracks tool number. NOT USED.
   int32_t line_number;          // Last line number sent
 
   float position[N_AXIS];       // Where the interpreter considers the tool to be at this point in the code
@@ -194,7 +194,7 @@ typedef struct {
 //   uint16_t command_words;  // NOTE: If this bitflag variable fills, G and M words can be separated.
 //   uint16_t value_words;
 
-  uint8_t non_modal_command;
+  uint32_t non_modal_command;
   gc_modal_t modal;
   gc_values_t values;
