+++ grbl/system.h
@@ -86,24 +86,24 @@
 
 // Define global system variables
 typedef struct {
-  uint8_t abort;                 // System abort flag. Forces exit back to main loop for reset.
-  uint8_t state;                 // Tracks the current state of Grbl.
-  uint8_t suspend;               // System suspend bitflag variable that manages holds, cancels, and safety door.
-  uint8_t soft_limit;            // Tracks soft limit errors for the state machine (Boolean)
-  uint8_t step_control;          // Regulates the step segment generator based on state.
+  uint32_t abort;                 // System abort flag. Forces exit back to main loop for reset.
+  uint32_t state;                 // Tracks the current state of Grbl.
+  uint32_t suspend;               // System suspend bitflag variable that manages holds, cancels, and safety door.
+  uint32_t soft_limit;            // Tracks soft limit errors for the state machine (Boolean)
+  uint32_t step_control;          // Regulates the step segment generator based on state.
 
   int32_t position[N_AXIS];      // Real-time machine (aka home) position vector in steps. 
                                  // NOTE: This may need to be a volatile variable, if problems arise.                             
 
   int32_t probe_position[N_AXIS]; // Last probe position in machine coordinates and steps.
-  uint8_t probe_succeeded;        // Tracks if last probing cycle was successful.
-  uint8_t homing_axis_lock;       // Locks axes when limits engage. Used as an axis motion mask in the stepper ISR.
+  uint32_t probe_succeeded;        // Tracks if last probing cycle was successful.
+  uint32_t homing_axis_lock;       // Locks axes when limits engage. Used as an axis motion mask in the stepper ISR.
 } system_t;
 extern system_t sys;
 
-extern volatile uint8_t sys_probe_state;    // Probing state value.  Used to coordinate the probing cycle with stepper ISR.
-extern volatile uint8_t sys_rt_exec_state;  // Global realtime executor bitflag variable for state management. See EXEC bitmasks.
-extern volatile uint8_t sys_rt_exec_alarm;  // Global realtime executor bitflag variable for setting various alarms.
+extern volatile uint32_t sys_probe_state;    // Probing state value.  Used to coordinate the probing cycle with stepper ISR.
+extern volatile uint32_t sys_rt_exec_state;  // Global realtime executor bitflag variable for state management. See EXEC bitmasks.
+extern volatile uint32_t sys_rt_exec_alarm;  // Global realtime executor bitflag variable for setting various alarms.
 
 // Initialize the serial protocol
 void system_init();
